+++ src/main/java/io/github/sakurawald/module/mixin/chat/history/SendChatHistoryMixin.java
@@ -12,10 +12,10 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(value = PlayerManager.class, priority = 1000 - 1)
-public class PlayerManagerMixin {
+public class SendChatHistoryMixin {
 
     @Inject(at = @At(value = "TAIL"), method = "onPlayerConnect")
-    private void sendChatHistoryToNewJoinedPlayer(ClientConnection connection, @NotNull ServerPlayerEntity serverPlayer, ConnectedClientData commonListenerCookie, CallbackInfo ci) {
+    void sendChatHistoryToNewJoinedPlayer(ClientConnection connection, @NotNull ServerPlayerEntity serverPlayer, ConnectedClientData commonListenerCookie, CallbackInfo ci) {
         ChatHistoryInitializer.getChatHistory().forEach(serverPlayer::sendMessage);
     }