+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (gameplay.carpet.fake_player_manager module) fix: we should always treat the fake-player name lower-case, to keep consistency with the vanilla `usercache.json` file and `carpet` mod. ([270](https://github.com/sakurawald/fuji/issues/270) Thanks to @Amotassic)
+- (chat.history module) fix: should not dupe the history entry. (Thanks to @Amotassic)